# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "2.0.0",
    "title": "Helpoo Api Documentation",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module, for <b>Helpoo</b>."
  },
  "servers": [
    {
      "url": "http://apidev.helpooapp.net/"
    },
    {
      "url": "https://apidev.helpooapp.net/"
    }
  ],
...
  "paths": {
...
    "/api/v2/settings/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "update settings",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "PhoneNumbers": {
                  "type": "string",
                  "example": "0123456789"
                },
                "email": {
                  "type": "string",
                  "example": "test@email.com"
                },
                "hotline": {
                  "type": "string",
                  "example": "19700"
                },
                "about": {
                  "type": "string",
                  "example": "about us"
                },
                "WALink": {
                  "type": "string",
                  "example": "https://wa.me/20123456789"
                }
              },
              "required": [
                "PhoneNumbers",
                "email",
                "hotline",
                "about",
                "WALink"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/settings/encodeString": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "decoded": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/settings/config/{id}": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v2/settings/callGETApi": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "example": "any"
                  },
                  "isMap": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/settings/wizard": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "create wizard",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "textAr": {
                  "type": "string",
                  "example": "مرحبا بكم في موقعنا"
                },
                "textEn": {
                  "type": "string",
                  "example": "Welcome to our website"
                },
                "img": {
                  "type": "string",
                  "example": "Base64 image"
                },
                "name": {
                  "type": "string",
                  "example": "welcome"
                },
                "headlineAr": {
                  "type": "string",
                  "example": "مرحبا بكم"
                },
                "headlineEn": {
                  "type": "string",
                  "example": "Welcome"
                },
                "apiKey": {
                  "type": "string",
                  "example": "123456789"
                },
                "time": {
                  "type": "string",
                  "example": "10"
                },
                "timeAfter": {
                  "type": "string",
                  "example": "10"
                },
                "timeBefore": {
                  "type": "string",
                  "example": "10"
                }
              },
              "required": [
                "textAr",
                "textEn",
                "img",
                "name",
                "headlineAr",
                "headlineEn",
                "apiKey",
                "time",
                "timeAfter",
                "timeBefore"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/settings/wizard/getByName/{name}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/settings/wizard/getOne/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/settings/wizard/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "update wizard",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "textAr": {
                  "type": "string",
                  "example": "مرحبا بكم في موقعنا"
                },
                "textEn": {
                  "type": "string",
                  "example": "Welcome to our website"
                },
                "img": {
                  "type": "string",
                  "example": "Base64 image"
                },
                "name": {
                  "type": "string",
                  "example": "welcome"
                },
                "headlineAr": {
                  "type": "string",
                  "example": "مرحبا بكم"
                },
                "headlineEn": {
                  "type": "string",
                  "example": "Welcome"
                },
                "apiKey": {
                  "type": "string",
                  "example": "123456789"
                },
                "time": {
                  "type": "string",
                  "example": "10"
                },
                "timeAfter": {
                  "type": "string",
                  "example": "10"
                },
                "timeBefore": {
                  "type": "string",
                  "example": "10"
                }
              },
              "required": [
                "textAr",
                "textEn",
                "img",
                "name",
                "headlineAr",
                "headlineEn",
                "apiKey",
                "time",
                "timeAfter",
                "timeBefore"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/settings/stats": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "sDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/settings/promoStats": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "sDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v2/settings/targets/getOne/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/settings/targets/create": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "create target",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "month": {
                  "type": "string",
                  "example": "December"
                },
                "target": {
                  "type": "number",
                  "example": 1000
                },
                "revenue": {
                  "type": "number",
                  "example": 1000
                },
                "achieved": {
                  "type": "boolean",
                  "example": true
                },
                "achievedPercentage": {
                  "type": "number",
                  "example": 12.5
                },
                "year": {
                  "type": "number",
                  "example": 2021
                }
              },
              "required": [
                "month",
                "target",
                "revenue",
                "achieved",
                "achievedPercentage",
                "year"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/settings/targets/update/{id}": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "update target",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "month": {
                  "type": "string",
                  "example": "December"
                },
                "target": {
                  "type": "number",
                  "example": 1000
                },
                "revenue": {
                  "type": "number",
                  "example": 1000
                },
                "achieved": {
                  "type": "boolean",
                  "example": true
                },
                "achievedPercentage": {
                  "type": "number",
                  "example": 12.5
                },
                "year": {
                  "type": "number",
                  "example": 2021
                }
              },
              "required": [
                "month",
                "target",
                "revenue",
                "achieved",
                "achievedPercentage",
                "year"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/settings/targets/delete/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}